# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: THREADCAST TTRPG API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: characters
    description: Character management
  - name: dice
    description: Saved dice preferences
  - name: rolls
    description: Shared gameplay roll log
  - name: casts
    description: Server-authoritative character casts
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List all characters for current user
      responses:
        "200":
          description: List of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a new character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"

  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get a character by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found
    patch:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterUpdate"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "409":
          description: Character sheet is stale
        "428":
          description: Character data updates require an expected version
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dice/preferences:
    get:
      operationId: getDicePreferences
      tags: [dice]
      summary: Get saved dice preferences for the current user
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Saved dice preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DicePreferences"
        "401":
          description: Not authenticated
    put:
      operationId: putDicePreferences
      tags: [dice]
      summary: Replace saved dice preferences for the current user
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DicePreferences"
      responses:
        "200":
          description: Updated dice preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DicePreferences"
        "400":
          description: Invalid dice preferences
        "401":
          description: Not authenticated

  /rolls:
    get:
      operationId: listRolls
      tags: [rolls]
      summary: Recent shared gameplay rolls
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Recent rolls, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameplayRoll"
        "401":
          description: Not authenticated
    post:
      operationId: createRoll
      tags: [rolls]
      summary: Generate and record a gameplay roll
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameplayRollInput"
      responses:
        "201":
          description: Recorded server-generated roll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameplayRoll"
        "400":
          description: Invalid roll request
        "401":
          description: Not authenticated
        "403":
          description: Character does not belong to the player

  /casts:
    post:
      operationId: createCast
      tags: [casts]
      summary: Atomically cast using a saved character sheet
      description: Charges resources, resolves the roll and consequences, and replays the complete response for an identical user/request ID.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CastInput"
      responses:
        "201":
          description: Server-resolved cast result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CastResult"
        "400":
          description: Invalid cast request
        "401":
          description: Not authenticated
        "403":
          description: Character unavailable to this player


  /rolls/discord/status:
    get:
      operationId: getRollDiscordStatus
      tags: [rolls]
      summary: Weavekeeper-only Discord delivery status
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Delivery status without webhook credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RollDiscordStatus"
        "403":
          description: Weavekeeper access required

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  schemas:
    GameplayRollInput:
      type: object
      additionalProperties: false
      required: [requestId, title, category, mode, modifier, diceSides, diceCount, multiplier]
      properties:
        requestId: { type: string, format: uuid }
        characterId: { type: integer, minimum: 1 }
        title: { type: string, minLength: 1, maxLength: 160 }
        category: { type: string, enum: [check, mend, support, damage, table] }
        mode: { type: string, enum: [NORMAL, HARMONY, DISCORD] }
        modifier: { type: integer, minimum: -50, maximum: 50 }
        diceSides: { type: integer, enum: [0, 4, 6, 8, 10, 12, 20] }
        diceCount: { type: integer, minimum: 0, maximum: 2 }
        bonusDiceSides: { type: integer, enum: [4, 6, 8, 10, 12] }
        bonusDiceCount: { type: integer, minimum: 1, maximum: 2 }
        multiplier: { type: integer, minimum: 1, maximum: 20 }
        dc: { type: integer, minimum: 1, maximum: 40 }
    GameplayRoll:
      type: object
      required: [id, playerName, characterName, title, category, mode, diceSides, d1, extraDice, modifier, multiplier, finalDie, total, isBreak, isMisfire, outcome, diceName, diceColor, createdAt]
      properties:
        id: { type: integer }
        playerName: { type: string }
        characterName: { type: string }
        title: { type: string }
        category: { type: string, enum: [check, cast, weave, mend, support, damage, table] }
        mode: { type: string, enum: [NORMAL, HARMONY, DISCORD] }
        diceSides: { type: integer }
        d1: { type: integer }
        d2: { type: integer }
        extraDice:
          type: array
          items:
            type: object
            required: [sides, value]
            properties:
              sides: { type: integer }
              value: { type: integer }
        modifier: { type: integer }
        multiplier: { type: integer }
        finalDie: { type: integer }
        total: { type: integer }
        dc: { type: integer }
        isBreak: { type: boolean }
        isMisfire: { type: boolean }
        outcome: { type: string }
        diceName: { type: string }
        diceColor: { type: string }
        createdAt: { type: string }
    CastInput:
      type: object
      additionalProperties: false
      required: [requestId, characterId, kind, intent, components]
      properties:
        requestId: { type: string, format: uuid }
        characterId: { type: integer, minimum: 1 }
        kind: { type: string, enum: [cast, weave] }
        intent: { type: string, minLength: 1, maxLength: 500 }
        components:
          type: array
          minItems: 1
          maxItems: 4
          items:
            $ref: "#/components/schemas/CastComponent"
    CastComponent:
      type: object
      additionalProperties: false
      required: [string, powerLevel, mode]
      properties:
        string: { type: string, minLength: 1, maxLength: 100 }
        powerLevel: { type: integer, minimum: 1, maximum: 5 }
        mode: { type: string, minLength: 1, maxLength: 40 }
    CastTableEffect:
      type: object
      required: [name, description]
      properties:
        name: { type: string }
        description: { type: string }
        damage:
          type: object
          required: [sides, count]
          properties:
            sides: { type: integer, enum: [4, 6, 8, 10, 12] }
            count: { type: integer, minimum: 1, maximum: 4 }
        burnout: { type: integer }
        resetTension: { type: boolean }
        condition: { type: string }
    CastAftermath:
      type: object
      required: [tension, pool, safeLimit, cost, overflow]
      properties:
        tension: { type: integer }
        pool: { type: integer }
        safeLimit: { type: integer }
        cost: { type: integer }
        overflow: { type: boolean }
        table:
          type: object
          required: [kind, die, effect]
          properties:
            kind: { type: string, enum: [Mishap, Snapback] }
            die: { type: integer }
            effect: { $ref: "#/components/schemas/CastTableEffect" }
        additionalTable:
          type: object
          required: [kind, die, effect]
          properties:
            kind: { type: string, enum: [Snapback] }
            die: { type: integer }
            effect: { $ref: "#/components/schemas/CastTableEffect" }
            damage: { type: integer }
        damage: { type: integer }
        strain:
          type: object
          required: [die, total, dc, failed]
          properties:
            die: { type: integer }
            total: { type: integer }
            dc: { type: integer }
            failed: { type: boolean }
        warning: { type: string }
    CastResult:
      type: object
      required: [roll, character, aftermath]
      properties:
        roll: { $ref: "#/components/schemas/GameplayRoll" }
        character: { $ref: "#/components/schemas/Character" }
        aftermath: { $ref: "#/components/schemas/CastAftermath" }
    RollDiscordStatus:
      type: object
      required: [configured, valid, recentSent, recentFailed, recentPending]
      properties:
        configured: { type: boolean }
        valid: { type: boolean }
        recentSent: { type: integer }
        recentFailed: { type: integer }
        recentPending: { type: integer }
        lastFailureAt: { type: ["string", "null"] }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthUser:
      type: object
      required: [id, username, displayName, role]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [player, weavekeeper]

    Character:
      type: object
      required: [id, userId, name, level, data, isDraft, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        level:
          type: integer
        affinity:
          type: ["string", "null"]
        mode:
          type: ["string", "null"]
        data:
          type: object
          description: Full character sheet data as JSONB
          additionalProperties: true
        isDraft:
          type: boolean
        version:
          type: integer
          minimum: 0
        createdAt:
          type: string
        updatedAt:
          type: string

    CharacterInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        level:
          type: integer
        affinity:
          type: string
        mode:
          type: string
        data:
          type: object
          additionalProperties: true
        isDraft:
          type: boolean

    CharacterUpdate:
      type: object
      additionalProperties: false
      properties:
        expectedUpdatedAt:
          type: string
          format: date-time
        expectedVersion:
          type: integer
          minimum: 0
        name:
          type: string
        level:
          type: integer
        affinity:
          type: string
        mode:
          type: string
        data:
          type: object
          additionalProperties: true
        isDraft:
          type: boolean

    DiceStyle:
      type: object
      additionalProperties: false
      required: [id, name, bodyColor, inkColor, edgeColor, finish, motif]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 40
        bodyColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        inkColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        edgeColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        finish:
          type: string
          enum: [matte, polished, glass]
        motif:
          type: string
          enum: [plain, weave, stars, etched]

    DicePreferences:
      type: object
      additionalProperties: false
      required: [sets, selectedId]
      properties:
        sets:
          type: array
          maxItems: 12
          items:
            $ref: "#/components/schemas/DiceStyle"
        selectedId:
          type: [string, "null"]
          format: uuid