# lib/api-zod/src/generated/types/castAftermath.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREADCAST TTRPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CastAftermathAdditionalTable } from './castAftermathAdditionalTable';
import type { CastAftermathStrain } from './castAftermathStrain';
import type { CastAftermathTable } from './castAftermathTable';

export interface CastAftermath {
  tension: number;
  pool: number;
  safeLimit: number;
  cost: number;
  overflow: boolean;
  table?: CastAftermathTable;
  additionalTable?: CastAftermathAdditionalTable;
  damage?: number;
  strain?: CastAftermathStrain;
  warning?: string;
}