# lib/api-zod/src/generated/types/castComponent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREADCAST TTRPG API
 * OpenAPI spec version: 0.1.0
 */

export interface CastComponent {
  /**
     * @minLength 1
     * @maxLength 100
     */
  string: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  powerLevel: number;
  /**
     * @minLength 1
     * @maxLength 40
     */
  mode: string;
}