# lib/api-zod/src/generated/types/castResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREADCAST TTRPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CastAftermath } from './castAftermath';
import type { Character } from './character';
import type { GameplayRoll } from './gameplayRoll';

export interface CastResult {
  roll: GameplayRoll;
  character: Character;
  aftermath: CastAftermath;
}