# lib/api-zod/src/generated/types/castTableEffect.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREADCAST TTRPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CastTableEffectDamage } from './castTableEffectDamage';

export interface CastTableEffect {
  name: string;
  description: string;
  damage?: CastTableEffectDamage;
  burnout?: number;
  resetTension?: boolean;
  condition?: string;
}