# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREADCAST TTRPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CharacterData } from './characterData';

export interface Character {
  id: number;
  userId: number;
  name: string;
  level: number;
  /** @nullable */
  affinity?: string | null;
  /** @nullable */
  mode?: string | null;
  /** Full character sheet data as JSONB */
  data: CharacterData;
  isDraft: boolean;
  /** @minimum 0 */
  version: number;
  createdAt: string;
  updatedAt: string;
}