# lib/api-zod/src/generated/types/characterUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREADCAST TTRPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CharacterUpdateData } from './characterUpdateData';

export interface CharacterUpdate {
  expectedUpdatedAt?: Date;
  /** @minimum 0 */
  expectedVersion?: number;
  name?: string;
  level?: number;
  affinity?: string;
  mode?: string;
  data?: CharacterUpdateData;
  isDraft?: boolean;
}